parser file made some changing.

+++ inc/parser.h
@@ -4,7 +4,7 @@
 int newFile(char *filename);
 int isWhiteChar(char c);
 int readWord(char *word);
-void fToStr(float val, int n, char *str);
+void fToStr(double val, int n, char *str);
 int parseToken(int idx, char* str, char* token, char delim);
 
 #endif
